Integrating makeall.sh inside setup_sm.sh so everything is automagically installed

+++ EventFilter/StorageManager/scripts/operations/mkstore.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# $Id: mkstore.sh,v 1.6 2008/06/11 13:58:04 loizides Exp $
+# $Id: mkstore.sh,v 1.7 2008/09/30 03:53:19 loizides Exp $
 
 if test -e "/etc/profile.d/sm_env.sh"; then 
     source /etc/profile.d/sm_env.sh
@@ -12,23 +12,24 @@ fi
 
 for i in emulator global; do
     cd $store 
-    mkdir -p $i
-    chmod 755 $i
-    find $i -type l -maxdepth 1 -exec rm -f "{}" \;
-    cd $store/$i;
-    mkdir -p mbox && chmod 777 mbox
-    mkdir -p log && chmod 777 log
-    rm -rf scripts
-    mkdir -p scripts && chmod 755 scripts
-done
+    if [ ! -d $i ]; then
+        mkdir -p $i
+        chmod 755 $i
+        find $i -type l -maxdepth 1 -exec rm -f "{}" \;
+        cd $store/$i;
+        mkdir -p mbox && chmod 777 mbox
+        mkdir -p log && chmod 777 log
+        rm -rf scripts
+        mkdir -p scripts && chmod 755 scripts
 
-for i in emulator global; do
-    cd $store/$i/scripts
-    touch dummy.pl
-    chmod 755 dummy.pl
-    for k in  insertFile.pl notifyTier0.pl closeFile.pl; do
-        ln -fs dummy.pl $k;
-    done
+        # Dealing with scripts
+        cd $store/$i/scripts
+        touch dummy.pl
+        chmod 755 dummy.pl
+        for k in  insertFile.pl notifyTier0.pl closeFile.pl; do
+            ln -fs dummy.pl $k;
+        done
+    fi
 done
 
 set counter=0
